chore: refactor image build
Refactoring image build to its own file, identical to vfkit but this time
I also fixed a couple lint issues. (and fixed bad merge from 2 min ago :( )
